requirements.txt
torch
huggingface_hub
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
unstructured
beautifulsoup4
requests
tqdm
python-dotenv
langchain_community
assemblyai
pypdf
bs4